+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inspecto-dev/types
 
+## 0.3.18
+
+### Patch Changes
+
+- release 0.3.18
+
 ## 0.3.17
 
 ### Patch Changes